Move inline helpers to their own file.
